# Convert RsaPaddingProcessor to statics