This is a follow up to #68553. This changes `RsaPaddingProcessor` to be entirely static.

* Instances are no longer cached as statics
* The DigestInfos can now use the `ReadOnlySpan<byte>` C# compiler optimization for arrays.

This has no material impact on benchmarks, in wall time nor per-invocation allocations, but makes the code a little more reasonable, and probably slightly reduces the amount of memory since the statics are no longer around indefinitely.


<details>
<summary>RSA Encrypt Benchmarks (click to expand)</summary>

|      Method |        Job | Toolchain | EncryptionPadding |        Mean |    Error |   StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------- |---------- |------------------ |------------:|---------:|---------:|------:|------:|------:|------:|----------:|
| EncryptData | Job-BWYKBU |    branch |        OaepSHA256 |    56.60 us | 0.078 us | 0.069 us |  1.00 |     - |     - |     - |     160 B |
| EncryptData | Job-JLGOQC |      main |        OaepSHA256 |    56.55 us | 0.064 us | 0.054 us |  1.00 |     - |     - |     - |     160 B |
|             |            |           |                   |             |          |          |       |       |       |       |           |
| DecryptData | Job-BWYKBU |    branch |        OaepSHA256 | 1,366.78 us | 0.584 us | 0.487 us |  1.00 |     - |     - |     - |     162 B |
| DecryptData | Job-JLGOQC |      main |        OaepSHA256 | 1,367.31 us | 1.072 us | 0.895 us |  1.00 |     - |     - |     - |     162 B |
|             |            |           |                   |             |          |          |       |       |       |       |           |
| EncryptData | Job-BWYKBU |    branch |             Pkcs1 |    55.84 us | 0.116 us | 0.108 us |  1.00 |     - |     - |     - |     160 B |
| EncryptData | Job-JLGOQC |      main |             Pkcs1 |    56.00 us | 0.165 us | 0.154 us |  1.00 |     - |     - |     - |     160 B |
|             |            |           |                   |             |          |          |       |       |       |       |           |
| DecryptData | Job-BWYKBU |    branch |             Pkcs1 | 1,365.71 us | 0.477 us | 0.398 us |  1.00 |     - |     - |     - |     162 B |
| DecryptData | Job-JLGOQC |      main |             Pkcs1 | 1,365.54 us | 0.818 us | 0.766 us |  1.00 |     - |     - |     - |     162 B |

</details>

<details>
<summary>RSA Sign Benchmarks (click to expand)</summary>

|     Method |        Job | Toolchain | SignaturePadding |        Mean |    Error |   StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------- |---------- |----------------- |------------:|---------:|---------:|------:|------:|------:|------:|----------:|
|   SignHash | Job-ENFCWV |    branch |            Pkcs1 | 1,371.64 μs | 0.529 μs | 0.442 μs |  1.00 |     - |     - |     - |      66 B |
|   SignHash | Job-UJMZPE |      main |            Pkcs1 | 1,372.17 μs | 0.783 μs | 0.694 μs |  1.00 |     - |     - |     - |      66 B |
|            |            |           |                  |             |          |          |       |       |       |       |           |
| VerifyHash | Job-ENFCWV |    branch |            Pkcs1 |    55.63 μs | 0.106 μs | 0.089 μs |  1.00 |     - |     - |     - |      32 B |
| VerifyHash | Job-UJMZPE |      main |            Pkcs1 |    55.63 μs | 0.135 μs | 0.113 μs |  1.00 |     - |     - |     - |      32 B |
|            |            |           |                  |             |          |          |       |       |       |       |           |
|   SignHash | Job-ENFCWV |    branch |              Pss | 1,367.65 μs | 0.474 μs | 0.396 μs |  1.00 |     - |     - |     - |     218 B |
|   SignHash | Job-UJMZPE |      main |              Pss | 1,368.40 μs | 0.984 μs | 0.872 μs |  1.00 |     - |     - |     - |     218 B |
|            |            |           |                  |             |          |          |       |       |       |       |           |
| VerifyHash | Job-ENFCWV |    branch |              Pss |    57.73 μs | 0.082 μs | 0.072 μs |  1.00 |     - |     - |     - |     216 B |
| VerifyHash | Job-UJMZPE |      main |              Pss |    57.70 μs | 0.133 μs | 0.125 μs |  1.00 |     - |     - |     - |     216 B |

</details>